@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_app/pages/login_page.dart';
 
 class HomePage extends StatefulWidget {
-  const HomePage({Key key}) : super(key: key);
+  const HomePage({Key? key}) : super(key: key);
 
   @override
   _HomePageState createState() => _HomePageState();
